+++ packages/stoa-core/src/gas/index.ts
@@ -1,3 +1,4 @@
 // @stoachain/stoa-core/gas
 
 export * from "./gasUtils.js";
+export * from "./yinEngine.js";